[m0][mx] Fix test_override_config_table failed in m0/mx (#8368)
What is the motivation for this PR?
test_override_config_table failed in m0/mx because default_pfcwd_status is enabled.

How did you do it?
Disable pfcwd before test and restore it after test.

How did you verify/test it?
Run tests on m0/mx/t1 topo, all pass.
